+++ src_vs2013/VSNDK.Package/UpdateManager/SimulatorManager.xaml.cs
@@ -72,7 +72,14 @@ namespace RIM.VSNDK_Package.UpdateManager
             }
             else
             {
-                umData.InstallAPI(((SimulatorsClass)((StackPanel)((Button)sender).Parent).DataContext).TargetVersion, false, true);
+                if (!GlobalFunctions.isOnline())
+                {
+                    System.Windows.MessageBox.Show("You are currently experiencing internet connection issues and cannot access the Update Manager server.  Please check your connection or try again later.", "Settings", MessageBoxButton.OK, MessageBoxImage.Warning, MessageBoxResult.OK);
+                }
+                else
+                {
+                    umData.InstallAPI(((SimulatorsClass)((StackPanel)((Button)sender).Parent).DataContext).TargetVersion, false, true);
+                }
             }
         }
 
@@ -89,7 +96,14 @@ namespace RIM.VSNDK_Package.UpdateManager
             }
             else
             {
-                umData.UninstallAPI(((SimulatorsClass)((StackPanel)((Button)sender).Parent).DataContext).TargetVersion, true);
+                if (!GlobalFunctions.isOnline())
+                {
+                    System.Windows.MessageBox.Show("You are currently experiencing internet connection issues and cannot access the Update Manager server.  Please check your connection or try again later.", "Settings", MessageBoxButton.OK, MessageBoxImage.Warning, MessageBoxResult.OK);
+                }
+                else
+                {
+                    umData.UninstallAPI(((SimulatorsClass)((StackPanel)((Button)sender).Parent).DataContext).TargetVersion, true);
+                }
             }
         }
